+++ R/data.R
@@ -1,4 +1,4 @@
-#' Kaya identity data for many countries
+#' Kaya identity data for many countries and regions
 #'
 #' A dataset containing Kaya identity parameters
 #' P, G, E, F, g, e, f, and ef for many countries
@@ -24,13 +24,14 @@
 #'   \item{G_mer}{Gross domestic product at market-exchange-rate,
 #'                in trillions of constant 2010 U.S. dollars}
 #' }
-#' @source \url{https://data.worldbank.org/indicator/SP.POP.TOTL},
-#' \url{https://data.worldbank.org/indicator/NY.GDP.MKTP.KD}, and
-#' \url{https://www.bp.com/en/global/corporate/energy-economics/statistical-review-of-world-energy/downloads.html}
+#' @source <https://data.worldbank.org/indicator/SP.POP.TOTL>,
+#' <https://data.worldbank.org/indicator/NY.GDP.MKTP.KD>, and
+#' <https://www.bp.com/en/global/corporate/energy-economics/statistical-review-of-world-energy/downloads.html>
 "kaya_data"
 
 
-#' Mix of fuels contributing to primary energy supply for many countries
+#' Mix of fuels contributing to primary energy supply for many countries and
+#' regions
 #'
 #' A dataset containing the fuel mix of how many quads and what fraction of
 #' total primary energy supply comes from coal, gas, oil, nuclear, and
@@ -49,7 +50,7 @@
 #'   \item{frac}{The fraction of that country or region's total primary
 #'              energy consumption from the fuel}
 #' }
-#' @source \url{https://www.bp.com/en/global/corporate/energy-economics/statistical-review-of-world-energy/downloads.html}
+#' @source <https://www.bp.com/en/global/corporate/energy-economics/statistical-review-of-world-energy/downloads.html>
 "fuel_mix"
 
 #' Top-down projections of future Kaya variables for many countries and regions
@@ -74,7 +75,7 @@
 #'   \item{ef}{Emissions intensity of the economy, in metric tons per
 #'             million dollars of GDP.}
 #' }
-#' @source \url{https://www.eia.gov/outlooks/archive/ieo17/}
+#' @source <https://www.eia.gov/outlooks/archive/ieo17/>
 "td_values"
 
 #' Top-down projections of trends in Kaya variables for many countries and regions
@@ -96,5 +97,5 @@
 #'   \item{f}{Trend in emissions intensity of the energy supply, in fraction per year}
 #'   \item{ef}{Trend in emissions intensity of the economy, in fraction per year}
 #' }
-#' @source \url{https://www.eia.gov/outlooks/archive/ieo17/}
+#' @source <https://www.eia.gov/outlooks/archive/ieo17/>
 "td_trends"