+++ R/kaya.R
@@ -2,6 +2,10 @@
 #' @keywords internal
 mtoe <- 1 / 25.2 # quads
 
+#' Conversion factor: quads per EJ
+#' @keywords internal
+EJ <- 0.947817120 # quads
+
 globalVariables(c("fuel_mix", "kaya_data", "region", "region_code",
                   "geography", "year", "P", "G", "E", "F", "g", "e", "f", "ef",
                   "td_values", "td_trends"))
@@ -10,7 +14,7 @@ globalVariables(c("fuel_mix", "kaya_data", "region", "region_code",
 #'
 #' @param region_code The three-letter country or region code
 #' @param data Data frame in which to look up `region_code`
-#' @param quiet       Suppress warnings if there is no such country or region.
+#' @param quiet Suppress warnings if there is no such country or region.
 #'
 #' @return The corresponding country or region name, or NULL if there is no
 #'         such country or region
@@ -42,16 +46,17 @@ kaya_region_list <- function() {
     as.character()
 }
 
-#' Get Kaya data for a country or region
+#' Get Kaya data for one or more countries or regions
 #'
-#' @param region_name The name of a country or region to look up
+#' @param region_name The name of one or more countries or regions to look up
 #' @param gdp         Use market exchange rates (`MER`) or purchasing power
 #'                    parity (`PPP`). Default is `MER`.
 #' @param quiet       Suppress warnings if there is no such country or region.
-#' @param region_code Optional three-letter country or region code to look up
+#' @param region_code Optional three-letter country or region codes to look up
 #'                     instead of the `region_name`
 #'
-#' @return a tibble of Kaya identity data for the country or region:
+#' @return a tibble of Kaya identity data for the countries or regions
+#'   specified:
 #' \describe{
 #'   \item{region}{The name of the country or region}
 #'   \item{year}{The year}
@@ -70,21 +75,22 @@ kaya_region_list <- function() {
 #'             million dollars of GDP.}
 #' }
 #'
-#' @details Units for G, g, e, and ef depend on whether the data is requested
-#'          in MER or PPP dollars: For MER, dollars are constant 2010 U.S.
+#' @details Units for _G_, _g_, _e_, and _ef_ depend on whether the data is
+#'          requested in MER or PPP dollars:
+#'          For MER, dollars are constant 2010 U.S.
 #'          dollars. For PPP, dollars are constant 2011 international dollars.
 #'
-#'          P and MER values for GDP and related quantities are available from
-#'          1960 onward.
+#'          _P_ and MER values for GDP and related quantities are available
+#'          from 1960 onward.
 #'
 #'          PPP values for GDP and related quantities are only available from
 #'          1990 onward.
 #'
-#'          Energy-related values (E, F, and derived quantities) are
+#'          Energy-related values (_E_, _F_, and derived quantities) are
 #'          available from 1965 onward.
 #'
-#'          Note that emissions (F, f, and ef) use millions of metric tons of
-#'          carbon dioxide, not carbon.
+#'          Note that emissions (_F_, _f_, and _ef_) are reported as millions
+#'          of metric tons of carbon dioxide, not carbon.
 #'
 #' @examples
 #' get_kaya_data("Brazil")
@@ -105,13 +111,11 @@ get_kaya_data <- function(region_name, gdp = c("MER", "PPP"), quiet = FALSE,
   data <- kayadata::kaya_data %>%
     dplyr::select(-region_code, -geography) %>%
     dplyr::filter(region %in% region_name)
-  if (nrow(data) == 0) {
-    if (!quiet) {
-      warning("There is no data for country or region ",
-              str_c(
-                ifelse(isTRUE(region_name == ""), region_code, region_name),
-                collapse = ", "))
-    }
+  if (nrow(data) == 0 && ! quiet) {
+    warning("There is no data for country or region ",
+            str_c(
+              ifelse(isTRUE(region_name == ""), region_code, region_name),
+              collapse = ", "))
   }
   if (gdp == "PPP") {
     data <- data %>% mutate(G = .data$G_ppp, g = G / P, e = E / G, ef = F / G)
@@ -121,19 +125,30 @@ get_kaya_data <- function(region_name, gdp = c("MER", "PPP"), quiet = FALSE,
   data
 }
 
-#' Get fuel mix for a country or region
+#' Get fuel mix for one or more countries or regions
 #'
-#' @param region_name The name of a country or region to look up
-#' @param collapse_renewables Combine hydroelectricity and other renewables into
-#'   a single category.
-#' @param quiet       Suppress warnings if there is no data for that country or
+#' @param region_name A character vector with the names of one or more
+#'   countries or regions to look up
+#' @param collapse_renewables Combine hydroelectricity and other renewables
+#'   into a single category.
+#' @param quiet Suppress warnings if there is no data for that country or
 #'   region.
-#' @param region_code Optional three-letter country or region code to look up
+#' @param region_code Optional three-letter country or region codes to look up
 #'   instead of the `region_name`
 #'
-#' @return a tibble of fuel mix for the country or region.
+#' @return A tibble of fuel mix for the countries or regions specified.
 #'   That is, the number of quads of each fuel and the
-#'   fraction of total primary energy coming from that fuel.
+#'   fraction of total primary energy coming from that fuel for each country
+#'   or region:
+#' \describe{
+#'   \item{region}{The name of the country or region}
+#'   \item{year}{The year reported}
+#'   \item{fuel}{The name of the fuel}
+#'   \item{quads}{The number of quads per year the country or region consumes}
+#'   \item{frac}{The fraction of the country's energy that comes from that
+#'     fuel}
+#' }
+#'
 #' @examples
 #' get_fuel_mix("United States")
 #' get_fuel_mix("World", collapse_renewables = FALSE)
@@ -163,7 +178,7 @@ get_fuel_mix <- function(region_name, collapse_renewables = TRUE,
       summarize_at(vars(quads, frac), list(~sum(., na.rm = T))) %>%
       ungroup()
   }
-  if (nrow(data) == 0 && is.null(region_code)) {
+  if (nrow(data) == 0 && ! quiet) {
     warning("There is no data for country or region ",
             str_c(
               ifelse(isTRUE(region_name == ""), region_code, region_name),
@@ -172,18 +187,18 @@ get_fuel_mix <- function(region_name, collapse_renewables = TRUE,
   data %>% arrange(region, fuel)
 }
 
-#' Get top-down trends for Kaya variables for a country or region, using
-#' projections from U.S. Energy Information Administration's International
-#' Energy Outlook report.
+#' Get top-down trends for Kaya variables for one or more countries or
+#' regions, using projections from U.S. Energy Information Administration's
+#' International Energy Outlook report.
 #'
-#' @param region_name The name of a country or region to look up
-#' @param quiet       Suppress warnings if there is no data for that country or
-#'                    region.
-#' @param region_code Optional three-letter country or region code to look up
-#'                     instead of the `region_name`
+#' @param region_name The name of one or more countries or regions to look up
+#' @param quiet Suppress warnings if there is no data for the specified
+#'   countries or regions.
+#' @param region_code Optional three-letter country or region codes to look up
+#'                    instead of the `region_name`
 #'
-#' @return a tibble of trends for P, G, E, F, g, e, f, and ef for the country,
-#' or region in percent per year.
+#' @return a tibble of trends for _P_, _G_, _E_, _F_, _g_, _e_, _f_, and _ef_
+#'   for each country or region in percent per year.
 #' @examples
 #' get_top_down_trends("Spain")
 #' get_top_down_trends(region_code = "RUS")
@@ -201,18 +216,17 @@ get_top_down_trends <- function(region_name, quiet = FALSE,
     dplyr::filter(region %in% region_name) %>%
     dplyr::mutate(g = G - P, e = E - G, f = F - E, ef = F - G) %>%
     dplyr::select(region, P, G, g, E, F, e, f, ef)
-  if (nrow(data) == 0 && is.null(region_code)) {
-    if (!quiet) {
-      warning("There is no data for country or region ",
-              str_c(
-                ifelse(isTRUE(region_name == ""), region_code, region_name),
-                collapse = ", "))
-    }
+  if (nrow(data) == 0 && !quiet) {
+    warning("There is no data for country or region ",
+            str_c(
+              ifelse(isTRUE(region_name == ""), region_code, region_name),
+              collapse = ", "))
   }
   data
 }
 
-#' Get top-down projections of Kaya variables for a country or region
+#' Get top-down projections of Kaya variables for one or more countries or
+#' regions
 #'
 #' @param region_name The name of a country or region to look up
 #' @param quiet       Suppress warnings if there is no data for that country or
@@ -220,8 +234,8 @@ get_top_down_trends <- function(region_name, quiet = FALSE,
 #' @param region_code Optional three-letter country or region code to look up
 #'                     instead of the `region_name`
 #'
-#' @return a tibble of values for P, G, E, F, g, e, f, and ef for the country
-#' or region:
+#' @return a tibble of values for _P_, _G_, _E_, _F_, _g_, _e_, _f_, and _ef_
+#'   for each country or region:
 #' \describe{
 #'   \item{region}{The name of the country or region}
 #'   \item{P}{Population, in billions}
@@ -258,30 +272,28 @@ get_top_down_values <- function(region_name, quiet = FALSE,
     dplyr::filter(region %in% region_name) %>%
     dplyr::mutate(g = G / P, e = E / G, f = F / E, ef = F / G) %>%
     dplyr::select(region, year, P, G, g, E, F, e, f, ef)
-  if (nrow(data) == 0 && is.null(region_code)) {
-    if (!quiet) {
-      warning("There is no data for country or region ",
-              str_c(
-                ifelse(isTRUE(region_name == ""), region_code, region_name),
-                collapse = ", "))
-    }
+  if (nrow(data) == 0 && !quiet) {
+    warning("There is no data for country or region ",
+            str_c(
+              ifelse(isTRUE(region_name == ""), region_code, region_name),
+              collapse = ", "))
   }
   data
 }
 
-#' Get top-down projections of Kaya variables for a country or region for a
-#' given year
+#' Get top-down projections of Kaya variables for one or more countries
+#' or regions for a given year
 #'
 #' @param region_name The name of a country or region to look up
-#' @param quiet       Suppress warnings if there is no data for that country or
-#'                    region.
+#' @param quiet Suppress warnings if there is no data for that country or
+#'              region.
 #' @param region_code Optional three-letter country or region code to look up
-#'                     instead of the `region_name`
+#'                    instead of the `region_name`
 #'
 #' @param year The year to project to
 #'
-#' @return a tibble of values for P, G, E, F, g, e, f, and ef for the country
-#' or region:
+#' @return a tibble of values for _P_, _G_, _E_, _F_, _g_, _e_, _f_, and _ef_
+#'   for each country or region:
 #' \describe{
 #'   \item{region}{The name of the country or region}
 #'   \item{year}{The year}
@@ -331,7 +343,10 @@ project_top_down <- function(region_name, year, quiet = FALSE,
     dplyr::select(region, year, P, G, g, E, F, e, f, ef)
   if (nrow(data) == 0 && is.null(region_code)) {
     if (!quiet) {
-      warning("There is no data for country or region ", region_name)
+      warning("There is no data for country or region ",
+              str_c(
+                ifelse(isTRUE(region_name == ""), region_code, region_name),
+                collapse = ", "))
     }
   }
   data
@@ -340,19 +355,24 @@ project_top_down <- function(region_name, year, quiet = FALSE,
 
 #' Get emission factors for different energy sources
 #'
+#' @param collapse_renewables Combine hydroelectricity and other renewables
+#'   into a single category.
 #' @return a tibble of values for emissions factors, in million metric
 #'         tons of carbon dioxide per quad of energy.
 #' @examples
 #' e_fac <- emissions_factors()
 #' e_fac
 #' @export
-emissions_factors <- function() {
-  tibble(
-    fuel = c("Coal", "Oil", "Natural Gas", "Nuclear", "Renewables"),
-    emission_factor = c(94.4, 70.0, 53.1, 0.0, 0.0)
+emissions_factors <- function(collapse_renewables = TRUE) {
+  ef <- tibble(
+    fuel = c("Coal", "Oil", "Natural Gas", "Nuclear", "Hydro", "Renewables"),
+    emission_factor = c(94.4, 70.0, 53.1, 0.0, 0.0, 0.0)
   ) %>%
-    mutate(fuel = ordered(fuel, levels = c("Coal", "Natural Gas", "Oil",
-                                           "Nuclear", "Renewables")))
+    mutate(fuel = ordered(fuel, levels = levels(fuel_mix$fuel)))
+  if (collapse_renewables) {
+    ef <- ef %>% filter(fuel != "Hydro") %>%
+      mutate(fuel = forcats::fct_recode(fuel, Renewables = "Hydro"))
+  }
 }
 
 #' Get power output from generation sources
@@ -377,9 +397,9 @@ emissions_factors <- function() {
 #' @references
 #' Environmental Protection Agency (2018) "Electric Power Monthly,"
 #' (October 2018) <https://www.eia.gov/electricity/monthly/current_month/epm.pdf>,
-#' \href{Table 6.7.A}{https://www.eia.gov/electricity/monthly/epm_table_grapher.php?t=epmt_6_07_a}.
+#' [Table 6.7.A](https://www.eia.gov/electricity/monthly/epm_table_grapher.php?t=epmt_6_07_a).
 #'
-#' Pielke, Jr., Roger A., \emph{The Climate Fix} (Basic Books, 2010).
+#' Pielke, Jr., Roger A., _The Climate Fix_ (Basic Books, 2010).
 #' @export
 generation_capacity <- function() {
   tibble(